requirements.txt
matplotlib==3.1.2
numpy==1.17.4
opencv_python==4.5.4.58
Pillow==9.1.0
scikit_image==0.16.2
scipy==1.3.3
skimage==0.0
